+++ HISTORY.md
@@ -3,6 +3,7 @@
 * Added support for downloading offer PDFs
 * Removed CLI
 * Replaced Maven with Gradle for building
+* Added support for recurring cycle numbers
 
 ## 1.0.4 (2013-04-13)